+++ CHANGELOG.md
@@ -7,11 +7,15 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+### Changed
+
+- `itmdump` no longer depends on the `mkfifo` command.
+
 ## v0.1.0 - 2016-10-03
 
 ### Added
 
-- `itmdump` tool that parses instrumentation packets from the stimulus port 0 and dumps the payload
-  to `stdout`. 
+- `itmdump` tool that parses instrumentation packets from the stimulus port 0
+  and dumps the payload to `stdout`. 
 
 [Unreleased]: https://github.com/japaric/itm/compare/v0.1.0...HEAD